renamed tmDbContext, reworked model relations using Fluent API
